- Add C wrappers for collision functions: moveWithCollisions, checkCollisions   - Add C wrappers for query functions: querySpritesAtPoint, querySpritesInRect,     querySpritesAlongLine, querySpriteInfoAlongLine, overlappingSprites,     allOverlappingSprites   - Add pd_sprite_freeArray helper to free collision/query result arrays   - Implement Go bindings with proper struct parsing for SpriteCollisionInfo     and SpriteQueryInfo   - Add SpriteQueryInfo type with entry/exit points for raycasting   - Fix CollisionVector type to use int instead of float32 (matches C API)   - Add sprite_collisions example demonstrating bouncing sprites, collision     response, and raycasting with QuerySpriteInfoAlongLine   - Update build_all.sh to include sprite_collisions example

+++ common.go
@@ -133,6 +133,6 @@ type CollisionPoint struct {
 
 // CollisionVector represents a vector in collision detection
 type CollisionVector struct {
-	X float32
-	Y float32
+	X int
+	Y int
 }
